rename: unified folder structure
